// frontend/src/client/types.gen.ts
export type ModelInfo = {
  /**
   * Id
   */
  id: string
  /**
   * Name
   */
  name: string
  /**
   * Description
   */
  description: string
  /**
   * Workflow Type
   */
  workflow_type?: 'streaming' | 'buffered'
  /**
   * Expected Latency Ms
   */
  expected_latency_ms?: [number, number]
}